1-arrat_iterator

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -10,12 +10,12 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	size_t i;
 
 	if (array == NULL || action == NULL)
 		return;
 
-	for (i = 0; i < (int) size; i++)
+	for (i = 0; i < size; i++)
 	{
 		action(array[i]);
 	}
